requirements.txt
alembic==1.10.2
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
elastic-transport==8.4.0
elasticsearch==8.7.0
Flask==2.2.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pika==1.3.2
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
zipp==3.15.0
